@@ -41,18 +47,22 @@
     "debounce": "^1.1.0",
     "fecha": "^2.3.3",
     "geval": "^2.2.0",
+    "gh-pages": "^2.1.1",
     "global": "^4.3.2",
     "history": "^4.7.2",
     "joi-browser": "^13.4.0",
     "localforage": "^1.7.1",
-    "material-ui-pickers": "^1.0.0-rc.10",
+    "mapbox-gl": "^1.4.0",
+    "material-ui-pickers": "^1.0.1",
+    "mathjs": "^6.2.1",
+    "memoize-one": "^5.1.1",
     "obstruction": "^2.1.0",
-    "prop-types": "^15.6.2",
+    "prop-types": "^15.7.2",
     "query-string": "^5.0.1",
     "raf": "^3.4.0",
     "raven-js": "^3.26.3",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "~16.5",
+    "react-dom": "~16.5",
     "react-error-boundary": "^1.2.2",
     "react-extras": "^0.7.1",
     "react-map-gl": "^3.2.10",
