@@ -2,30 +2,35 @@ import { push } from 'connected-react-router';
 import document from 'global/document';
 import * as Types from './types';
 import Timelineworker from '../timeline';
+import { getDongleID } from '../url';
 
-export function updateState (data) {
+export function updateState(data) {
   return {
     type: Types.WORKER_STATE_UPDATE,
-    data: data
+    data
   };
 }
 
-export function selectRange (start, end) {
+export function selectRange(start, end) {
   return (dispatch, getState) => {
     const state = getState();
     if (!state.workerState.dongleId) {
-      return dispatch({
+      dispatch({
         type: Types.TIMELINE_SELECTION_CHANGED,
-        start, end
+        start,
+        end
       });
+      return;
     }
     const curPath = document.location.pathname;
-    var desiredPath = urlForState(state.workerState.dongleId, start, end);
+    const dongleId = getDongleID(curPath) || state.workerState.dongleId;
+    const desiredPath = urlForState(dongleId, start, end);
 
     if (state.zoom.start !== start || state.zoom.end !== end) {
       dispatch({
         type: Types.TIMELINE_SELECTION_CHANGED,
-        start, end
+        start,
+        end
       });
     }
 
@@ -43,14 +48,14 @@ export function selectRange (start, end) {
   };
 }
 
-export function selectDevice (dongleId) {
+export function selectDevice(dongleId) {
   return (dispatch, getState) => {
     const state = getState();
     if (state.workerState.dongleId !== dongleId) {
       Timelineworker.selectDevice(dongleId);
     }
     const curPath = document.location.pathname;
-    var desiredPath = urlForState(dongleId, state.zoom.start, state.zoom.end);
+    const desiredPath = urlForState(dongleId, state.zoom.start, state.zoom.end);
 
     if (curPath !== desiredPath) {
       dispatch(push(desiredPath));
@@ -58,13 +63,13 @@ export function selectDevice (dongleId) {
   };
 }
 
-function urlForState (dongleId, start, end) {
-  var path = [dongleId];
+function urlForState(dongleId, start, end) {
+  const path = [dongleId];
 
   if (start && end) {
     path.push(start);
     path.push(end);
   }
 
-  return '/' + path.join('/');
+  return `/${path.join('/')}`;
 }
