@@ -1,4 +1,4 @@
-let Colors = {
+const Colors = {
   transparent: 'transparent',
   white: '#fff',
   white10: 'rgba(255, 255, 255, 0.1)',
