@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import Obstruction from 'obstruction';
 import { partial } from 'ap';
 import fecha from 'fecha';
-import { classNames } from 'react-extras';
 
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
@@ -13,55 +12,48 @@ import Badge from '@material-ui/core/Badge';
 import AnnotationList from './AnnotationList';
 import { filterEvent } from '../../utils';
 
-const styles = theme => {
-  return {
-    root: {
-    },
-    upsellDemo: {
-      cursor: 'default',
-      pointerEvents: 'none',
-      opacity: 0.8
-    },
-    annotationsViewer: {
-      display: 'flex',
-      position: 'absolute',
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-      overflow: 'hidden',
-      flexDirection: 'column',
-    },
-    annotationsViewerTabs: {
-      marginBottom: 12,
-      width: '100%',
-    },
-    tabLabelBadge: {
-      backgroundColor: theme.palette.grey[100],
-      borderRadius: 24,
-      marginRight: 24,
-      minWidth: 28,
-      top: -4,
-    },
-    tabLabelText: {
-      fontWeight: 500,
-      paddingLeft: 24,
-      textTransform: 'none',
-    },
-    annotationsViewerTabIndicator: {
-      backgroundColor: theme.palette.grey[100],
-      borderRadius: 8,
-      height: 6,
-    },
-    annotationsViewerList: {
-      height: '100%',
-      overflowY: 'scroll',
-    },
-  };
-};
+const styles = (theme) => ({
+  root: {
+  },
+  annotationsViewer: {
+    display: 'flex',
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    overflow: 'hidden',
+    flexDirection: 'column',
+  },
+  annotationsViewerTabs: {
+    marginBottom: 12,
+    width: '100%',
+  },
+  tabLabelBadge: {
+    backgroundColor: theme.palette.grey[100],
+    borderRadius: 24,
+    marginRight: 24,
+    minWidth: 28,
+    top: -4,
+  },
+  tabLabelText: {
+    fontWeight: 500,
+    paddingLeft: 24,
+    textTransform: 'none',
+  },
+  annotationsViewerTabIndicator: {
+    backgroundColor: theme.palette.grey[100],
+    borderRadius: 8,
+    height: 6,
+  },
+  annotationsViewerList: {
+    height: '100%',
+    overflowY: 'scroll',
+  },
+});
 
 class AnnotationTabs extends Component {
-  constructor (props) {
+  constructor(props) {
     super(props);
 
     this.handleChange = this.handleChange.bind(this);
@@ -71,11 +63,11 @@ class AnnotationTabs extends Component {
     };
   }
 
-  handleChange (event, selectedTab) {
+  handleChange(event, selectedTab) {
     this.setState({ selectedTab });
   }
 
-  count (resolved) {
+  count(resolved) {
     return this.props.segment.events.filter((event) => {
       if (!filterEvent(event)) {
         return false;
@@ -86,41 +78,42 @@ class AnnotationTabs extends Component {
       if (!resolved && !event.id) {
         return true;
       }
-    }).length
+    }).length;
   }
 
   render() {
     const { classes } = this.props;
     return (
       <div style={{ position: 'relative', height: '100%' }}>
-        <div className={ classes.annotationsViewer }>
+        <div className={classes.annotationsViewer}>
           <Tabs
-            value={ this.state.selectedTab }
-            onChange={ this.handleChange }
+            value={this.state.selectedTab}
+            onChange={this.handleChange}
             classes={{
               indicator: classes.annotationsViewerTabIndicator
             }}
-            className={ classNames(classes.annotationsViewerTabs, {
-              [classes.upsellDemo]: this.props.isUpsellDemo
-            })}>
-            <Tab label={
-              <Typography className={ classes.tabLabelText }>
-                <Badge classes={{ badge: classes.tabLabelBadge }} badgeContent={ this.count(false) }>
+            className={classes.annotationsViewerTabs}
+          >
+            <Tab label={(
+              <Typography className={classes.tabLabelText}>
+                <Badge classes={{ badge: classes.tabLabelBadge }} badgeContent={this.count(false)}>
                   &nbsp;
                 </Badge>
                 Annotations Unresolved
               </Typography>
-            } />
-            <Tab label={
-              <Typography className={ classes.tabLabelText }>
-                <Badge classes={{ badge: classes.tabLabelBadge }} badgeContent={ this.count(true) }>
+            )}
+            />
+            <Tab label={(
+              <Typography className={classes.tabLabelText}>
+                <Badge classes={{ badge: classes.tabLabelBadge }} badgeContent={this.count(true)}>
                   &nbsp;
                 </Badge>
                 Resolved
               </Typography>
-            } />
+            )}
+            />
           </Tabs>
-          <div className={ classes.annotationsViewerList + ' x-scrollbar' }>
+          <div className={`${classes.annotationsViewerList} x-scrollbar`}>
             { this.renderTab(this.state.selectedTab) }
           </div>
         </div>
@@ -128,18 +121,17 @@ class AnnotationTabs extends Component {
     );
   }
 
-  renderTab (index) {
+  renderTab(index) {
     switch (index) {
       case 0:
         return (
           <AnnotationList
-            segment={ this.props.segment }
+            segment={this.props.segment}
             unresolved
-            isUpsellDemo={ this.props.isUpsellDemo }
           />
         );
       case 1:
-        return (<AnnotationList segment={ this.props.segment } resolved />);
+        return (<AnnotationList segment={this.props.segment} resolved />);
     }
   }
 }
