@@ -16,7 +16,7 @@ import {
 import DriveMap from '../DriveMap';
 import DriveVideo from '../DriveVideo';
 
-const styles = theme => ({
+const styles = (theme) => ({
   mediaOptions: {
     alignItems: 'center',
     border: '1px solid rgba(255,255,255,.1)',
@@ -71,10 +71,9 @@ const MediaType = {
   DRIVER_VIDEO: 'dcamera',
   HUD: 'hud',
   MAP: 'map'
-}
+};
 
 class Media extends Component {
-
   constructor(props) {
     super(props);
 
@@ -82,87 +81,97 @@ class Media extends Component {
 
     this.state = {
       inView: MediaType.HUD,
-    }
+    };
   }
 
-  render () {
+  render() {
     const { classes } = this.props;
-    let { inView } = this.state;
+    const { inView } = this.state;
     return (
-      <React.Fragment>
+      <>
         { this.renderMediaOptions() }
         { inView === MediaType.MAP && <DriveMap /> }
-        { (inView !== MediaType.MAP) &&
+        { (inView !== MediaType.MAP)
+          && (
           <DriveVideo
-            shouldShowUI={ inView === MediaType.HUD }
-            front={ inView === MediaType.DRIVER_VIDEO }
-            onVideoChange={ (noVideo) => {
-              this.setState({ inView: noVideo ? MediaType.MAP : inView }) }
-            } />
-        }
-      </React.Fragment>
+            shouldShowUI={inView === MediaType.HUD}
+            front={inView === MediaType.DRIVER_VIDEO}
+            onVideoChange={(noVideo) => {
+              this.setState({ inView: noVideo ? MediaType.MAP : inView });
+            }}
+          />
+          )}
+      </>
     );
   }
 
-  renderMediaOptions () {
+  renderMediaOptions() {
     const { classes, currentSegment } = this.props;
-    let { inView } = this.state;
+    const { inView } = this.state;
     const mediaSource = 'eon-road-camera';
     const hasDriverCameraStream = this.props.currentSegment && this.props.currentSegment.hasDriverCameraStream;
     return (
       <Grid container>
-        <Grid item xs={ 5 }>
-          {/*
-          <FormControl className={ classes.mediaSource }>
-            <Select
-              name='media-source'
-              value={ 'eon-road-camera' }
-              className={ classes.mediaSourceSelect }>
-              <MenuItem value='eon-road-camera'>EON Road Camera</MenuItem>
-              <MenuItem value='eon-driver-camera'>EON Driver Camera</MenuItem>
-            </Select>
-          </FormControl>
-          */}
-        </Grid>
-        <Grid item xs={ 7 }
-          className={ classes.mediaOptions }>
-          <Grid item xs={ 3 }
-            className={ classes.mediaOption }
-            style={ inView === MediaType.HUD ? { opacity: 1 } : { } }
-            onClick={() => this.setState({inView: MediaType.HUD }) }>
-            <Typography className={ classes.mediaOptionText }>
+        <Grid
+          item
+          xs={hasDriverCameraStream ? 5 : 7}
+        />
+        <Grid
+          item
+          xs={hasDriverCameraStream ? 7 : 5}
+          className={classes.mediaOptions}
+        >
+          <Grid
+            item
+            xs={hasDriverCameraStream ? 3 : 4}
+            className={classes.mediaOption}
+            style={inView === MediaType.HUD ? { opacity: 1 } : { }}
+            onClick={() => this.setState({ inView: MediaType.HUD })}
+          >
+            <Typography className={classes.mediaOptionText}>
               HUD
             </Typography>
           </Grid>
-          <Grid item xs={ 3 }
-            className={ classes.mediaOption }
-            style={ inView === MediaType.VIDEO ? { opacity: 1 } : {}}
-            onClick={() => this.setState({inView: MediaType.VIDEO})}>
-            <Typography className={ classes.mediaOptionText }>
+          <Grid
+            item
+            xs={hasDriverCameraStream ? 3 : 4}
+            className={classes.mediaOption}
+            style={inView === MediaType.VIDEO ? { opacity: 1 } : {}}
+            onClick={() => this.setState({ inView: MediaType.VIDEO })}
+          >
+            <Typography className={classes.mediaOptionText}>
               Video
             </Typography>
           </Grid>
-          <Grid item xs={ 3 }
-            className={ cx(classes.mediaOption, { disabled: !hasDriverCameraStream } ) }
-            style={ inView === MediaType.DRIVER_VIDEO ? { opacity: 1 } : {} }
-            onClick={() => hasDriverCameraStream && this.setState({inView: MediaType.DRIVER_VIDEO})}>
-            <Typography className={ classes.mediaOptionText }>
-              Driver Video
-            </Typography>
-          </Grid>
-          <Grid item xs={ 3 }
-            className={ classes.mediaOption }
-            style={ inView === MediaType.MAP ? { opacity: 1 } : { } }
-            onClick={() => this.setState({inView: MediaType.MAP})}>
-            <Typography className={ classes.mediaOptionText }>
+          { hasDriverCameraStream
+            && (
+            <Grid
+              item
+              xs={3}
+              className={cx(classes.mediaOption, { disabled: !hasDriverCameraStream })}
+              style={inView === MediaType.DRIVER_VIDEO ? { opacity: 1 } : {}}
+              onClick={() => hasDriverCameraStream && this.setState({ inView: MediaType.DRIVER_VIDEO })}
+            >
+              <Typography className={classes.mediaOptionText}>
+                Driver Video
+              </Typography>
+            </Grid>
+            )}
+          <Grid
+            item
+            xs={hasDriverCameraStream ? 3 : 4}
+            className={classes.mediaOption}
+            style={inView === MediaType.MAP ? { opacity: 1 } : { }}
+            onClick={() => this.setState({ inView: MediaType.MAP })}
+          >
+            <Typography className={classes.mediaOptionText}>
               Map
             </Typography>
           </Grid>
         </Grid>
       </Grid>
-    )
+    );
   }
-
 }
 
 const stateToProps = Obstruction({
