@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import Obstruction from 'obstruction';
+import { classNames } from 'react-extras';
 import PropTypes from 'prop-types';
 import document from 'global/document';
 import { timeout } from 'thyming';
@@ -13,71 +14,50 @@ import Snackbar from '@material-ui/core/Snackbar';
 
 import { raw as RawApi } from '@commaai/comma-api';
 import TimelineWorker from '../../timeline';
+import * as Demo from '../../demo';
+const demoFiles = require('../../demo/files.json');
 
-const styles = theme => {
-  return {
-    root: {
-      width: '100%',
-      height: '100%',
-      paddingTop: theme.spacing.unit,
-      background: 'transparent',
-      textAlign: 'right'
-    },
-    footerButton: {
-      display: 'inline-block',
-      border: '1px solid ' + theme.palette.grey[800],
-      color: theme.palette.grey[50],
-      textDecoration: 'none',
-      borderRadius: 20,
-      padding: '10px 20px',
-      margin: '0px 10px',
-      cursor: 'pointer'
-    }
+const styles = (theme) => ({
+  root: {
+    width: '100%',
+    height: '100%',
+    paddingTop: theme.spacing.unit,
+    background: 'transparent',
+    textAlign: 'right'
+  },
+  footerButton: {
+    display: 'inline-block',
+    border: `1px solid ${theme.palette.grey[800]}`,
+    color: theme.palette.grey[50],
+    textDecoration: 'none',
+    borderRadius: 20,
+    padding: '10px 20px',
+    margin: '0px 10px',
+    cursor: 'pointer'
   }
-};
+});
 
 class AnnotationsFooter extends Component {
-  static propTypes = {
-    segment: PropTypes.object.isRequired,
-  };
-
-  constructor (props) {
+  constructor(props) {
     super(props);
 
     this.copySegmentName = this.copySegmentName.bind(this);
     this.handleClose = this.handleClose.bind(this);
     this.downloadSegmentFile = this.downloadSegmentFile.bind(this);
     this.openInCabana = this.openInCabana.bind(this);
-    this.openWayInOsmEditor = this.openWayInOsmEditor.bind(this);
-    this.updateWayId = this.updateWayId.bind(this);
 
     this.state = {
       showCopiedSnack: false,
-      wayId: 0,
+    };
+  }
+
+  copySegmentName(e) {
+    const { segment } = this.props;
+    if (!segment) {
+      return;
     }
-  }
-
-  componentDidMount() {
-    raf(this.updateWayId);
-  }
-
-  updateWayId() {
-    raf(this.updateWayId);
-
-    var event = TimelineWorker.currentLiveMapData();
-    if (event && event.LiveMapData) {
-      var wayId = parseInt(event.LiveMapData.WayId);
-      if (wayId !== this.state.wayId) {
-        this.setState({ wayId });
-      }
-    } else if (this.state.wayId) {
-      this.setState({ wayId: 0 });
-    }
-  }
-
-  copySegmentName (e) {
     const el = document.createElement('textarea');
-    el.value = this.props.segment.route + '--' + this.props.segment.segment;
+    el.value = `${segment.route}--${segment.segment}`;
     el.setAttribute('readonly', '');
     el.style.position = 'absolute';
     el.style.left = '-9999px';
@@ -97,75 +77,75 @@ class AnnotationsFooter extends Component {
       return;
     }
 
-    let seg = this.props.segment;
-    let segmentKeyPath = seg.route.replace('|', '/') + '/' + seg.segment;
+    const seg = this.props.segment;
+    const segmentKeyPath = `${seg.route.replace('|', '/')}/${seg.segment}`;
 
-    let files = (await RawApi.getRouteFiles(this.props.segment.route));
-    let url = files[type].find(function(url) {
-      return url.indexOf(segmentKeyPath) !== -1;
-    })
+    let files;
+    if (Demo.isDemo()) {
+      files = demoFiles;
+    } else {
+      files = (await RawApi.getRouteFiles(this.props.segment.route));
+    }
+    const url = files[type].find((url) => url.indexOf(segmentKeyPath) !== -1);
 
     if (url) {
       window.location.href = url;
     }
   }
 
-  handleClose () {
+  handleClose() {
     this.setState({
       showCopiedSnack: false
     });
     this.snackTimer();
   }
 
-  openInCabana () {
-    var params = {
-      'route': this.props.segment.route,
-      'url': this.props.segment.url,
-      'seekTime': Math.floor((TimelineWorker.currentOffset() - this.props.segment.routeOffset) / 1000)
+  openInCabana() {
+    const { segment, loop, start } = this.props;
+    const currentOffset = TimelineWorker.currentOffset();
+    const params = {
+      route: segment.route,
+      url: segment.url,
+      seekTime: Math.floor((currentOffset - segment.routeOffset) / 1000)
     };
-    var win = window.open('https://my.comma.ai/cabana/' + qs.stringify(params, true) , '_blank');
+    const routeStartTime = (start + segment.routeOffset);
+
+    if (loop.startTime && loop.startTime > routeStartTime && loop.duration < 180000) {
+      const startTime = Math.floor((loop.startTime - routeStartTime) / 1000);
+      params.segments = [startTime, Math.floor(startTime + (loop.duration / 1000))].join(',');
+    }
+
+    const win = window.open(`https://my.comma.ai/cabana/${qs.stringify(params, true)}`, '_blank');
     if (win.focus) {
       win.focus();
     }
   }
 
-  openWayInOsmEditor () {
-    var { wayId } = this.state;
-    if (!wayId) {
-      return;
-    }
-
-    var win = window.open(`https://www.openstreetmap.org/way/${wayId}`, '_blank');
-    if (win.focus) {
-      win.focus();
-    }
-  }
-
-  render () {
+  render() {
     return (
-      <Paper className={ this.props.classes.root } >
-        { this.props.segment && this.props.segment.hasVideo &&
-          <a className={ this.props.classes.footerButton } onClick={ partial(this.downloadSegmentFile, 'cameras') }>
+      <Paper className={this.props.classes.root}>
+        { this.props.segment && this.props.segment.hasVideo
+          && (
+          <a className={this.props.classes.footerButton} onClick={partial(this.downloadSegmentFile, 'cameras')}>
             Download Camera Segment
           </a>
-        }
-        { this.props.segment && this.props.segment.hasDriverCamera &&
-          <a className={ this.props.classes.footerButton } onClick={ partial(this.downloadSegmentFile, 'dcameras') }>
+          )}
+        { this.props.segment && this.props.segment.hasDriverCamera
+          && (
+          <a className={this.props.classes.footerButton} onClick={partial(this.downloadSegmentFile, 'dcameras')}>
             Download Front Camera Segment
           </a>
-        }
-        { this.props.segment &&
-          <a className={ this.props.classes.footerButton } onClick={ partial(this.downloadSegmentFile, 'logs') }>
+          )}
+        { this.props.segment
+          && (
+          <a className={this.props.classes.footerButton} onClick={partial(this.downloadSegmentFile, 'logs')}>
             Download Log Segment
           </a>
-        }
-        <a className={ this.props.classes.footerButton } onClick={ this.openWayInOsmEditor } href='#'>
-          Edit OpenStreetMap Here
-        </a>
-        <a className={ this.props.classes.footerButton } onClick={ this.openInCabana } href='#'>
+          )}
+        <a className={classNames(this.props.classes.footerButton, 'openInCabana')} onClick={this.openInCabana} href="#">
           View CAN Data in Cabana
         </a>
-        <a className={ this.props.classes.footerButton } onClick={ this.copySegmentName } target='_blank'>
+        <a className={this.props.classes.footerButton} onClick={this.copySegmentName} target="_blank">
           Copy Current Segment
         </a>
         <Snackbar
@@ -185,5 +165,11 @@ class AnnotationsFooter extends Component {
     );
   }
 }
+
+AnnotationsFooter.propTypes = {
+  segment: PropTypes.object.isRequired,
+  loop: PropTypes.object.isRequired,
+  start: PropTypes.number.isRequired
+};
 
 export default withStyles(styles)(AnnotationsFooter);
