@@ -21,60 +21,58 @@ import Pencil from '@material-ui/icons/Edit';
 
 import { devices as Devices } from '@commaai/comma-api';
 import Timelineworker from '../../timeline';
-import EonUpsell from '../Annotations/eonUpsell';
+import CommaTwoUpsell from '../Annotations/commaTwoUpsell';
 
-const styles = theme => {
-  return {
-    base: {
-      height: '100%',
-      overflowY: 'scroll',
-    },
-    device: {
-      alignItems: 'center',
-      cursor: 'pointer',
-      display: 'flex',
-      padding: 32,
-      paddingTop: 16,
-      paddingBottom: 16,
-      '&.isSelected': {
-        backgroundColor: '#171B1D',
-      }
-    },
-    deviceAvatar: {
-      backgroundColor: '#1D2225',
-      borderRadius: 30,
-      height: 46,
-      width: 46,
-    },
-    deviceInfo: {
-      display: 'flex',
-      flexDirection: 'column',
-      justifyContent: 'center',
-      marginLeft: 16,
-    },
-    deviceAlias: {},
-    deviceId: {
-      color: '#525E66',
-      fontFamily: 'MaisonNeueMono',
-    },
+const styles = (theme) => ({
+  base: {
+    height: '100%',
+    overflowY: 'scroll',
+  },
+  device: {
+    alignItems: 'center',
+    cursor: 'pointer',
+    display: 'flex',
+    padding: 32,
+    paddingTop: 16,
+    paddingBottom: 16,
+    '&.isSelected': {
+      backgroundColor: '#171B1D',
+    }
+  },
+  deviceAvatar: {
+    backgroundColor: '#1D2225',
+    borderRadius: 30,
+    height: 46,
+    width: 46,
+  },
+  deviceInfo: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    marginLeft: 16,
+  },
+  deviceAlias: {},
+  deviceId: {
+    color: '#525E66',
+    fontFamily: 'MaisonNeueMono',
+  },
 
-    editDeviceIcon: {
-      color: 'white',
-      '&:hover': {
-        color: theme.palette.grey[100]
-      }
-    },
-    nameField: {
-      marginRight: theme.spacing.unit,
-    },
-    saveButton: {
-      marginRight: theme.spacing.unit,
-    },
-    textField: {
-      marginBottom: theme.spacing.unit
-    },
-  }
-};
+  editDeviceIcon: {
+    color: 'white',
+    '&:hover': {
+      color: theme.palette.grey[100]
+    }
+  },
+  nameField: {
+    marginRight: theme.spacing.unit,
+  },
+  saveButton: {
+    marginRight: theme.spacing.unit,
+  },
+  textField: {
+    marginBottom: theme.spacing.unit
+  },
+});
 
 class DeviceList extends Component {
   constructor(props) {
@@ -102,7 +100,7 @@ class DeviceList extends Component {
     }
   }
 
-  toggleDeviceEdit (device) {
+  toggleDeviceEdit(device) {
     if (this.state.editingDevice === device.dongle_id) {
       this.setState({ editingDevice: null });
     } else {
@@ -110,136 +108,145 @@ class DeviceList extends Component {
       this.setState({ editingDevice: device.dongle_id, deviceAlias: device.alias });
     }
   }
-  cancelEdit () {
+
+  cancelEdit() {
     this.setState({ editingDevice: null });
   }
 
-  handleAliasChange (e) {
+  handleAliasChange(e) {
     this.setState({ deviceAlias: e.target.value });
   }
 
-  handleAliasFieldKeyPress (dongle_id, e) {
+  handleAliasFieldKeyPress(dongle_id, e) {
     if (e.key === 'Enter' && !this.state.isWaitingForApi) {
       this.setDeviceAlias(dongle_id);
     }
   }
 
-  async setDeviceAlias (dongle_id) {
+  async setDeviceAlias(dongle_id) {
     this.setState({ isWaitingForApi: true });
     try {
       const device = await Devices.setDeviceAlias(dongle_id, this.state.deviceAlias.trim());
       Timelineworker.updateDevice(device);
       this.setState({ isWaitingForApi: false, editingDevice: null });
-    } catch(e) {
+    } catch (e) {
       Raven.captureException(e);
       this.setState({ error: e.message, isWaitingForApi: false });
     }
   }
-  deviceTypePretty (deviceType) {
-    if (deviceType === 'neo'){
+
+  deviceTypePretty(deviceType) {
+    if (deviceType === 'neo') {
       return 'EON';
+    } else if (deviceType === 'two') {
+      return 'comma two';
     }
     return deviceType;
   }
-  render () {
-    var devices = this.props.devices;
-    var dongleId = this.props.selectedDevice;
-    var found = !dongleId;
-    var onlyHasAppDevice = true;
-    devices.forEach(function (device, idx) {
-      if (device.dongle_id === dongleId) {
-        found = true;
-      }
-      onlyHasAppDevice = onlyHasAppDevice && (device.device_type !== 'neo' && device.device_type !== 'panda');
-    });
+
+  render() {
+    let { devices } = this.props;
+    const dongleId = this.props.selectedDevice;
+    let found = devices.some((device) => device.dongle_id === dongleId);
+    let onlyHasAppDevice = (devices.length === 0);
 
     if (!found) {
-      devices.push({
+      devices = [{
         dongle_id: dongleId,
         shared: true,
         alias: 'Shared device',
-      });
+      }].concat(devices);
     }
 
     return (
-      <div className={ this.props.classes.base }>
+      <div className={this.props.classes.base}>
         { devices.filter(this.filterDrivingDevice).map(this.renderDevice) }
-        { onlyHasAppDevice && <EonUpsell hook='Upgrade to an EON to augment your driving experience' /> }
+        { onlyHasAppDevice && <CommaTwoUpsell hook="Get started with comma two" /> }
       </div>
     );
   }
 
-  renderDevice (device) {
+  renderDevice(device) {
     const { classes } = this.props;
     const isSelected = (this.props.selectedDevice === device.dongle_id);
-    let alias = device.alias || this.deviceTypePretty(device.device_type);
+    const alias = device.alias || this.deviceTypePretty(device.device_type);
     return (
       <div
-        key={ device.dongle_id }
-        onClick={ partial(this.props.handleDeviceSelected, device.dongle_id) }
-        className={ cx(classes.device, [{ isSelected: isSelected }]) }>
-        <div className={ classes.deviceAvatar } />
-        <div className={ classes.deviceInfo }>
-          <Typography variant='body2' className={ classes.deviceAlias }>
+        key={device.dongle_id}
+        onClick={partial(this.props.handleDeviceSelected, device.dongle_id)}
+        className={cx(classes.device, [{ isSelected }])}
+      >
+        <div className={classes.deviceAvatar} />
+        <div className={classes.deviceInfo}>
+          <Typography variant="body2" className={classes.deviceAlias}>
             { alias }
           </Typography>
-          <Typography variant='caption' className={ classes.deviceId }>
-            ({ device.dongle_id })
+          <Typography variant="caption" className={classes.deviceId}>
+            (
+            { device.dongle_id }
+)
           </Typography>
         </div>
       </div>
-    )
+    );
 
     const oldRender = (
       <ExpansionPanel
         classes={{ expanded: classes.expanded }}
-        key={ device.dongle_id }
-        expanded={ this.props.selectedDevice === device.dongle_id }
-        onChange={ partial(this.props.handleDeviceSelected, device.dongle_id) }
-        className={ classes.expansion } >
+        key={device.dongle_id}
+        expanded={this.props.selectedDevice === device.dongle_id}
+        onChange={partial(this.props.handleDeviceSelected, device.dongle_id)}
+        className={classes.expansion}
+      >
         <ExpansionPanelSummary>
           <Grid container>
-            <Grid item xs={ 10 }>
-              { this.state.editingDevice === device.dongle_id ?
-                <React.Fragment>
-                  { this.state.isWaitingForApi && <LinearProgress /> }
-                  { this.state.error !== null && <FormHelperText error>{ this.state.error }</FormHelperText> }
-                  <TextField
-                    id="name"
-                    label="Name"
-                    className={ classes.textField }
-                    value={ this.state.deviceAlias }
-                    onChange={ this.handleAliasChange }
-                    onKeyPress={ partial(this.handleAliasFieldKeyPress, device.dongle_id) } />
-                </React.Fragment>
-                :
-                <Typography className={ classes.deviceListItemName }>
-                  { (alias + ' (' + device.dongle_id + ')') }
-                </Typography>
-              }
+            <Grid item xs={10}>
+              { this.state.editingDevice === device.dongle_id
+                ? (
+                  <>
+                    { this.state.isWaitingForApi && <LinearProgress /> }
+                    { this.state.error !== null && <FormHelperText error>{ this.state.error }</FormHelperText> }
+                    <TextField
+                      id="name"
+                      label="Name"
+                      className={classes.textField}
+                      value={this.state.deviceAlias}
+                      onChange={this.handleAliasChange}
+                      onKeyPress={partial(this.handleAliasFieldKeyPress, device.dongle_id)}
+                    />
+                  </>
+                )
+                : (
+                  <Typography className={classes.deviceListItemName}>
+                    { (`${alias} (${device.dongle_id})`) }
+                  </Typography>
+                )}
             </Grid>
-            { (!device.shared && (device.is_owner || this.props.isSuperUser)) &&
-              <Grid item xs={ 2 }>
-                <Pencil className={ classes.editDeviceIcon } onClick={ partial(this.toggleDeviceEdit, device) } />
+            { (!device.shared && (device.is_owner || this.props.isSuperUser))
+              && (
+              <Grid item xs={2}>
+                <Pencil className={classes.editDeviceIcon} onClick={partial(this.toggleDeviceEdit, device)} />
               </Grid>
-            }
-            { this.state.editingDevice === device.dongle_id &&
-              <React.Fragment>
-                <Grid item xs={ 6 }>
-                  <Button
-                  variant='outlined'
-                  onClick={ partial(this.setDeviceAlias, device.dongle_id) }
-                  className={ classes.saveButton }>
+              )}
+            { this.state.editingDevice === device.dongle_id
+              && (
+                <>
+                  <Grid item xs={6}>
+                    <Button
+                      variant="outlined"
+                      onClick={partial(this.setDeviceAlias, device.dongle_id)}
+                      className={classes.saveButton}
+                    >
                     Save
-                  </Button>
-                </Grid>
-                <Grid item xs={ 6 }>
-                  <Button variant='outlined' onClick={ this.cancelEdit }>
+                    </Button>
+                  </Grid>
+                  <Grid item xs={6}>
+                    <Button variant="outlined" onClick={this.cancelEdit}>
                     Cancel
-                  </Button>
-                </Grid>
-              </React.Fragment>
-            }
+                    </Button>
+                  </Grid>
+                </>
+              )}
           </Grid>
         </ExpansionPanelSummary>
       </ExpansionPanel>
