@@ -22,45 +22,43 @@ import DriveListItem from './DriveListItem';
 
 const MIN_TIME_BETWEEN_ROUTES = 60000; // 1 minute
 
-const styles = theme => {
-  return {
-    header: {
-      alignItems: 'center',
-      borderBottom: '1px solid rgba(255, 255, 255, 0.1)',
-      padding: 16,
-      paddingLeft: 48,
-      paddingRight: 60,
-    },
-    headerLabel: {
-      cursor: 'default',
-      textTransform: 'uppercase',
-    },
-    drives: {
-      height: '100%',
-      margin: 0,
-      overflowY: 'scroll',
-      padding: 16,
-      paddingLeft: 24,
-      paddingRight: 24,
-    },
-    zeroState: {
-      padding: '16px 48px',
-    },
-    settingsArea: {
-      display: 'flex',
-      justifyContent: 'flex-end',
-    },
-    settingsButton: {
-      border: '1px solid #272D30'
-    },
-    settingsButtonIcon: {
-      color: '#272D30',
-    },
-  }
-};
+const styles = (theme) => ({
+  header: {
+    alignItems: 'center',
+    borderBottom: '1px solid rgba(255, 255, 255, 0.1)',
+    padding: 16,
+    paddingLeft: 48,
+    paddingRight: 60,
+  },
+  headerLabel: {
+    cursor: 'default',
+    textTransform: 'uppercase',
+  },
+  drives: {
+    height: '100%',
+    margin: 0,
+    overflowY: 'scroll',
+    padding: 16,
+    paddingLeft: 24,
+    paddingRight: 24,
+  },
+  zeroState: {
+    padding: '16px 48px',
+  },
+  settingsArea: {
+    display: 'flex',
+    justifyContent: 'flex-end',
+  },
+  settingsButton: {
+    border: '1px solid #272D30'
+  },
+  settingsButtonIcon: {
+    color: '#272D30',
+  },
+});
 
 class DriveList extends Component {
-  constructor (props) {
+  constructor(props) {
     super(props);
 
     this.goToAnnotation = this.goToAnnotation.bind(this);
@@ -69,8 +67,8 @@ class DriveList extends Component {
     this.renderDriveListSettings = this.renderDriveListSettings.bind(this);
     this.handleClickedSettings = this.handleClickedSettings.bind(this);
     this.handleClosedSettings = this.handleClosedSettings.bind(this);
-    this.handleOpenedSettingsModal= this.handleOpenedSettingsModal.bind(this);
-    this.handleCanceledSettings= this.handleCanceledSettings.bind(this);
+    this.handleOpenedSettingsModal = this.handleOpenedSettingsModal.bind(this);
+    this.handleCanceledSettings = this.handleCanceledSettings.bind(this);
     this.handleClosedSettingsModal = this.handleClosedSettingsModal.bind(this);
 
     this.state = {
@@ -79,58 +77,63 @@ class DriveList extends Component {
     };
   }
 
-  componentWillReceiveProps (props) {
+  componentWillReceiveProps(props) {
     if (props.device && !this.state.deviceAliasSaved) {
       this.setState({ deviceAliasSaved: props.device.alias });
     }
   }
 
-  filterShortDrives (ride) {
+  filterShortDrives(ride) {
     return ride.duration > 60000;
   }
 
-  goToAnnotation (segment) {
-    let startTime = segment.startTime - this.props.zoomBuffer;
-    let endTime = segment.startTime + segment.duration + this.props.zoomBuffer;
+  goToAnnotation(segment) {
+    const startTime = segment.startTime - this.props.zoomBuffer;
+    const endTime = segment.startTime + segment.duration + this.props.zoomBuffer;
     this.props.dispatch(selectRange(startTime, endTime));
   }
 
-  handleClickedSettings (event) {
+  handleClickedSettings(event) {
     this.setState({ anchorEl: event.currentTarget });
   }
 
-  handleClosedSettings () {
+  handleClosedSettings() {
     this.setState({ anchorEl: null });
   }
 
-  handleOpenedSettingsModal (device) {
-    this.setState({ showDeviceSettingsModal: true })
+  handleOpenedSettingsModal(device) {
+    this.setState({ showDeviceSettingsModal: true });
   }
 
   handleClosedSettingsModal() {
     this.setState({ showDeviceSettingsModal: false });
   }
 
-  handleCanceledSettings () {
+  handleCanceledSettings() {
     this.setState({
       anchorEl: null,
       showDeviceSettingsModal: false,
-    })
+    });
   }
 
-  handleClose () {
+  handleClose() {
     this.setState({
       showPicker: false
     });
   }
 
-  render () {
+  render() {
     const { classes, device, segments } = this.props;
-    var driveList = [];
-    var lastEnd = 0;
-    var lastSegmentEnd = 0;
-    var curRideChunk = null;
-    this.props.segments.forEach(function (segment) {
+
+    if (!device) {
+      return [];
+    }
+
+    const driveList = [];
+    let lastEnd = 0;
+    let lastSegmentEnd = 0;
+    let curRideChunk = null;
+    this.props.segments.forEach((segment) => {
       if (!curRideChunk || segment.startTime - lastEnd > MIN_TIME_BETWEEN_ROUTES) {
         curRideChunk = {
           segments: 0,
@@ -151,43 +154,45 @@ class DriveList extends Component {
       curRideChunk.segments++;
       lastEnd = segment.startTime + segment.duration;
       curRideChunk.annotations += segment.events.filter(filterEvent)
-        .reduce((memo, event) => event.id ? memo : memo + 1, 0);
+        .reduce((memo, event) => (event.id ? memo : memo + 1), 0);
     });
 
     return (
-      <React.Fragment>
+      <>
         { this.renderDriveListHeader() }
         { driveList.length === 0 && this.renderZeroRides() }
-        <ul className={ classes.drives }>
-          { driveList.filter(this.filterShortDrives).map((drive) => {
-              return (
-                <DriveListItem
-                  key={ drive.startTime }
-                  drive={ drive }
-                  deviceAlias={ device.alias } />
-              )
-            })
-          }
+        <ul className={classes.drives}>
+          { driveList.filter(this.filterShortDrives).map((drive) => (
+            <DriveListItem
+              key={drive.startTime}
+              drive={drive}
+              deviceAlias={device.alias}
+            />
+          ))}
         </ul>
-      </React.Fragment>
+      </>
     );
   }
 
   renderZeroRides() {
     const { classes } = this.props;
-    var zeroRidesEle = null;
-    let device = this.props.device;
-    let hasDrivesInQuery = (device && device.last_segment_utc_millis !== null)
-      && (device.last_segment_utc_millis >= this.props.start);
+    let zeroRidesEle = null;
+    const { device, segmentData } = this.props;
 
-    if (hasDrivesInQuery) {
+    if (device && (segmentData === null || typeof segmentData.segments === 'undefined')) {
       zeroRidesEle = <Typography>Loading...</Typography>;
-    } else {
-      zeroRidesEle = <Typography>Looks like you haven{'\''}t driven in the selected time range.</Typography>
+    } else if (segmentData && segmentData.segments && segmentData.segments.length === 0) {
+      zeroRidesEle = (
+        <Typography>
+Looks like you haven
+          {'\''}
+t driven in the selected time range.
+        </Typography>
+      );
     }
 
     return (
-      <div className={ classes.zeroState }>
+      <div className={classes.zeroState}>
         <Grid container>
           { zeroRidesEle }
         </Grid>
@@ -198,41 +203,42 @@ class DriveList extends Component {
   renderDriveListHeader() {
     const { classes, device } = this.props;
     return (
-      <div className={ classes.header }>
-        <Grid container alignItems='center'>
-          <Grid item xs={ 4 }>
-            <Typography variant='title'>
-              { device.alias } Drives
+      <div className={classes.header}>
+        <Grid container alignItems="center">
+          <Grid item xs={4}>
+            <Typography variant="title">
+              { device.alias }
+              {' '}
+Drives
             </Typography>
           </Grid>
-          <Grid item xs={ 2 }>
-            <Typography variant='caption' className={ classes.headerLabel }>
+          <Grid item xs={2}>
+            <Typography variant="caption" className={classes.headerLabel}>
               Duration
             </Typography>
           </Grid>
-          <Grid item xs={ 2 }>
-            <Typography variant='caption' className={ classes.headerLabel }>
+          <Grid item xs={2}>
+            <Typography variant="caption" className={classes.headerLabel}>
               Origin
             </Typography>
           </Grid>
-          <Grid item xs={ 2 }>
-            <Typography variant='caption' className={ classes.headerLabel }>
+          <Grid item xs={2}>
+            <Typography variant="caption" className={classes.headerLabel}>
               Destination
             </Typography>
           </Grid>
-          <Grid item xs={ 1 }>
-            <Typography variant='caption' className={ classes.headerLabel }>
+          <Grid item xs={1}>
+            <Typography variant="caption" className={classes.headerLabel}>
               Distance
             </Typography>
           </Grid>
-          <Grid item xs={ 1 } className={ classes.settingsArea }>
-            { ((device && !device.shared && device.is_owner) || this.props.isSuperUser) &&
-              this.renderDriveListSettings()
-            }
+          <Grid item xs={1} className={classes.settingsArea}>
+            { ((device && !device.shared && device.is_owner) || this.props.isSuperUser)
+              && this.renderDriveListSettings()}
           </Grid>
         </Grid>
       </div>
-    )
+    );
   }
 
   renderDriveListSettings() {
@@ -240,24 +246,26 @@ class DriveList extends Component {
     const { anchorEl } = this.state;
     const open = Boolean(anchorEl);
     return (
-      <React.Fragment>
+      <>
         <IconButton
-          className={ classes.settingsButton }
-          onClick={ partial(this.handleOpenedSettingsModal, device) }>
-          <SettingsIcon className={ classes.settingsButtonIcon } />
+          className={classes.settingsButton}
+          onClick={partial(this.handleOpenedSettingsModal, device)}
+        >
+          <SettingsIcon className={classes.settingsButtonIcon} />
         </IconButton>
         <DeviceSettingsModal
-          isOpen={ this.state.showDeviceSettingsModal }
-          onClose={ this.handleClosedSettingsModal }
-          onCancel={ this.handleCanceledSettings }
+          isOpen={this.state.showDeviceSettingsModal}
+          onClose={this.handleClosedSettingsModal}
+          onCancel={this.handleCanceledSettings}
         />
-      </React.Fragment>
-    )
+      </>
+    );
   }
 }
 
 const stateToProps = Obstruction({
   segments: 'workerState.segments',
+  segmentData: 'workerState.segmentData',
   start: 'workerState.start',
   device: 'workerState.device',
   isSuperUser: 'workerState.profile.superuser',
