@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import Obstruction from 'obstruction';
 import { partial } from 'ap';
 import fecha from 'fecha';
+import { classNames } from 'react-extras';
 
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
@@ -15,64 +16,62 @@ import Timeline from '../Timeline';
 import { RightArrow } from '../../icons';
 import { KM_PER_MI } from '../../utils/conversions';
 
-const styles = theme => {
-  return {
-    drive: {
-      background: 'rgba(255, 255, 255, 0.0)',
-      background: 'linear-gradient(to bottom, #30373B 0%, #1D2225 100%)',
-      borderTop: '1px solid rgba(255, 255, 255, .05)',
-      borderRadius: 8,
-      cursor: 'pointer',
-      display: 'flex',
-      flexDirection: 'column',
-      marginBottom: 12,
-      overflow: 'hidden',
-      padding: 0,
-      transition: 'background .2s',
-      '&:hover': {}
-    },
-    driveHeader: {
-      alignItems: 'center',
-      padding: 18,
-      paddingLeft: 24,
-      paddingRight: 24,
-    },
-    driveHeaderIntro: {
-      display: 'flex',
-    },
-    driveAvatar: {
-      alignItems: 'center',
-      background: '#404B4F',
-      borderRadius: 30,
-      color: '#fff',
-      display: 'flex',
-      fontWeight: 600,
-      height: 52,
-      justifyContent: 'center',
-      margin: '0px 3%',
-      width: 52,
-    },
-    driveTitle: {
-      marginLeft: '10%',
-    },
-    driveTimeline: {},
-    driveArrow: {
-      color: '#404B4F',
-      height: '100%',
-      marginLeft: '25%',
-      width: 32,
-    },
-  }
-};
+const styles = (theme) => ({
+  drive: {
+    background: 'rgba(255, 255, 255, 0.0)',
+    background: 'linear-gradient(to bottom, #30373B 0%, #1D2225 100%)',
+    borderTop: '1px solid rgba(255, 255, 255, .05)',
+    borderRadius: 8,
+    cursor: 'pointer',
+    display: 'flex',
+    flexDirection: 'column',
+    marginBottom: 12,
+    overflow: 'hidden',
+    padding: 0,
+    transition: 'background .2s',
+    '&:hover': {}
+  },
+  driveHeader: {
+    alignItems: 'center',
+    padding: 18,
+    paddingLeft: 24,
+    paddingRight: 24,
+  },
+  driveHeaderIntro: {
+    display: 'flex',
+  },
+  driveAvatar: {
+    alignItems: 'center',
+    background: '#404B4F',
+    borderRadius: 30,
+    color: '#fff',
+    display: 'flex',
+    fontWeight: 600,
+    height: 52,
+    justifyContent: 'center',
+    margin: '0px 3%',
+    width: 52,
+  },
+  driveTitle: {
+    marginLeft: '10%',
+  },
+  driveTimeline: {},
+  driveArrow: {
+    color: '#404B4F',
+    height: '100%',
+    marginLeft: '25%',
+    width: 32,
+  },
+});
 
 class DriveListDrive extends Component {
-  constructor (props) {
+  constructor(props) {
     super(props);
 
     this.state = {
       startLocation: null,
       endLocation: null,
-    }
+    };
 
     this.handleDriveClicked = this.handleDriveClicked.bind(this);
   }
@@ -84,101 +83,107 @@ class DriveListDrive extends Component {
       if (!this.mounted) {
         return;
       }
-      this.setState({ startLocation })
-    })
+      this.setState({ startLocation });
+    });
     GeocodeApi().reverseLookup(drive.endCoord).then((endLocation) => {
       if (!this.mounted) {
         return;
       }
-      this.setState({ endLocation })
-    })
+      this.setState({ endLocation });
+    });
   }
+
   componentWillUnmount() {
     this.mounted = false;
   }
-  handleDriveClicked (drive) {
-    let startTime = drive.startTime - 1000;
-    let endTime = drive.startTime + drive.duration + 1000;
+
+  handleDriveClicked(drive) {
+    const startTime = drive.startTime - 1000;
+    const endTime = drive.startTime + drive.duration + 1000;
     this.props.dispatch(selectRange(startTime, endTime));
   }
 
-  render () {
+  render() {
     const { drive, deviceAlias, classes } = this.props;
     const { startLocation, endLocation } = this.state;
     const startTime = fecha.format(new Date(drive.startTime), 'HH:mm');
     const startDate = fecha.format(new Date(drive.startTime), 'ddd, MMM D');
-    const endTime = fecha.format(new Date(drive.startTime + drive.duration + 1000), 'HH:mm')
+    const endTime = fecha.format(new Date(drive.startTime + drive.duration + 1000), 'HH:mm');
     const duration = formatDriveDuration(drive.duration);
     const points = getDrivePoints(drive.duration);
     return (
       <li
-        key={ drive.startTime }
-        className={ classes.drive }
-        onClick={ partial(this.handleDriveClicked, drive) }>
-        <div className={ classes.driveHeader }>
+        key={drive.startTime}
+        className={classNames(classes.drive, 'DriveEntry')}
+        onClick={partial(this.handleDriveClicked, drive)}
+      >
+        <div className={classes.driveHeader}>
           <Grid container>
-            <Grid item xs={ 4 } className={ classes.driveHeaderIntro }>
-              <div className={ classes.driveAvatar }>
+            <Grid item xs={4} className={classes.driveHeaderIntro}>
+              <div className={classes.driveAvatar}>
                 { drive.annotations }
               </div>
-              <div className={ classes.driveTitle }>
-                <Typography variant='body2'>
-                  { startDate } @ { startTime } to { endTime }
+              <div className={classes.driveTitle}>
+                <Typography variant="body2">
+                  { `${startDate} @ ${startTime} to ${endTime}` }
                 </Typography>
                 <Typography>
                   { deviceAlias }
                 </Typography>
               </div>
             </Grid>
-            <Grid item xs={ 2 }>
-              <Typography variant='body2'>
-                { duration.hours > 0 && duration.hours.toString() + 'hr ' }{ duration.minutes }min
+            <Grid item xs={2}>
+              <Typography variant="body2">
+                { duration.hours > 0 && `${duration.hours.toString()}hr ` }
+                { `${duration.minutes} min` }
               </Typography>
               <Typography>
-                { points } points
+                { `${points} points` }
               </Typography>
             </Grid>
-            <Grid item xs={ 2 }>
-              <Typography variant='body2'>
+            <Grid item xs={2}>
+              <Typography variant="body2">
                 { startLocation && startLocation.neighborhood }
               </Typography>
               <Typography>
-                { startLocation && (startLocation.locality + ', ' + startLocation.region) }
+                { startLocation && (`${startLocation.locality}, ${startLocation.region}`) }
               </Typography>
             </Grid>
-            <Grid item xs={ 2 }>
-              <Typography variant='body2'>
+            <Grid item xs={2}>
+              <Typography variant="body2">
                 { endLocation && endLocation.neighborhood }
               </Typography>
               <Typography>
-                { endLocation && (endLocation.locality + ', ' + endLocation.region) }
+                { endLocation && (`${endLocation.locality}, ${endLocation.region}`) }
               </Typography>
             </Grid>
-            <Grid item xs={ 1 }>
-              <Typography variant='body2'>
-                { +drive.distanceMiles.toFixed(1) } mi
+            <Grid item xs={1}>
+              <Typography variant="body2">
+                { `${+drive.distanceMiles.toFixed(1)} mi` }
               </Typography>
               <Typography>
-                { +(drive.distanceMiles * KM_PER_MI).toFixed(1) } km
+                { `${+(drive.distanceMiles * KM_PER_MI).toFixed(1)} km` }
               </Typography>
             </Grid>
-            <Grid item xs={ 1 }>
-              <RightArrow className={ classes.driveArrow } />
+            <Grid item xs={1}>
+              <RightArrow className={classes.driveArrow} />
             </Grid>
           </Grid>
         </div>
         <Timeline
-          className={ classes.driveTimeline }
-          zoomed colored hasThumbnails hasGradient
+          className={classes.driveTimeline}
+          zoomed
+          colored
+          hasThumbnails
+          hasGradient
           zoomOverride={{
             start: drive.startTime,
             end: drive.startTime + drive.duration
-          }} />
+          }}
+        />
       </li>
     );
   }
 }
 
-const stateToProps = Obstruction({});
-
-export default connect(stateToProps)(withStyles(styles)(DriveListDrive));
+export default connect(() => ({}))(withStyles(styles)(DriveListDrive));
