@@ -5,45 +5,92 @@ import Hls from '@commaai/hls.js';
 export default class HLSSource extends Component {
   constructor(props, context) {
     super(props, context);
-    this.hls = new Hls({disablePtsDtsCorrectionInMp4Remux: true});
+    this.hls = new Hls({
+      enableWorker: false,
+      disablePtsDtsCorrectionInMp4Remux: false
+    });
+    this.hls.on(Hls.Events.MANIFEST_PARSED, () => {
+      this.props.video.play();
+    });
+    this.hls.on(Hls.Events.BUFFER_APPENDED, (eventName, data) => {
+      if (this.props.onBufferAppend) {
+        this.props.onBufferAppend();
+      }
+    });
+
+    this.hls.on(Hls.Events.ERROR, (event, data) => {
+      if (data.fatal) {
+        switch (data.type) {
+          case Hls.ErrorTypes.NETWORK_ERROR:
+          // try to recover network error
+            console.log('fatal network error encountered, try to recover');
+            this.hls.startLoad();
+            break;
+          case Hls.ErrorTypes.MEDIA_ERROR:
+            console.log('fatal media error encountered, try to recover');
+            this.hls.recoverMediaError();
+            break;
+          default:
+          // cannot recover
+            this.hls.destroy();
+            this.hls = null;
+            break;
+        }
+      }
+    });
+
+    this.state = {
+      src: ''
+    };
+    // this.hls.on(Hls.Events.STREAM_STATE_TRANSITION, (eventName, data) => {
+    // });
   }
 
-  componentDidUpdate (prevProps, prevState) {
-    if (this.props.src !== prevProps.src || this.props.video !== prevProps.video) {
-      this.initHls();
+  componentDidMount() {
+    this.setState({
+      src: this.props.src
+    });
+  }
+
+  componentWillUpdate(nextProps, nextState) {
+    if (nextProps.src !== nextState.src) {
+      // console.log('Loading media source!', nextProps.src);
+      if (this.state.src && this.state.src.length) {
+        // console.log('this.hls.detachMedia();');
+        this.hls.detachMedia();
+      }
+      this.setState({
+        src: nextProps.src
+      });
     }
   }
-  componentWillMount() {
-    this.initHls();
-  }
 
-  initHls () {
-    // `src` is the property get from this component
-    // `video` is the property insert from `Video` component
-    // `video` is the html5 video element
-    const { src, video } = this.props;
-
-    // load hls video source base on hls.js
-    if (Hls.isSupported()) {
-      // if (this.hls) {
-      //   this.hls.destroy();
-      // }
-      // this.hls = new Hls();
+  componentDidUpdate(prevProps, prevState) {
+    const { src } = this.state;
+    if (src !== prevState.src && src && src.length) {
+      // console.log('this.hls.loadSource(src);');
       this.hls.loadSource(src);
-      this.hls.attachMedia(video);
+      // console.log('this.hls.attachMedia(this.props.video);');
+      this.hls.attachMedia(this.props.video);
+
+      if (this.props.onSourceLoaded) {
+        this.props.onSourceLoaded();
+      }
     }
   }
 
   componentWillUnmount() {
     // destroy hls video source
     if (this.hls) {
+      // console.log('this.hls.destroy();');
       this.hls.destroy();
+      this.hls = null;
     }
   }
 
   render() {
     return (
-      <source src={this.props.src} type={this.props.type || 'application/x-mpegURL'} />
+      <source src={this.state.src} type={this.props.type || 'application/x-mpegURL'} />
     );
   }
 }
