@@ -1,22 +1,20 @@
-import React, { Component } from 'react';
+import React, { PureComponent } from 'react';
 import { connect } from 'react-redux';
 import Obstruction from 'obstruction';
 
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 
-const styles = theme => {
-  return {
-    root: {}
-  }
-};
+const styles = (/* theme */) => ({
+  root: {}
+});
 
-class MyAwesomeComponent extends Component {
-  render () {
+class MyAwesomeComponent extends PureComponent {
+  render() {
     return (
-      <React.Fragment>
+      <>
         <Typography>asdf</Typography>
-      </React.Fragment>
+      </>
     );
   }
 }
