@@ -1,33 +1,33 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux'
+import { connect } from 'react-redux';
 import raf from 'raf';
+import { withStyles } from '@material-ui/core/styles';
+import PropTypes from 'prop-types';
 import TimelineWorker from '../timeline';
 
-import { withStyles } from '@material-ui/core/styles';
-
-const styles = theme => {
-  return {
-    root: {
-      ...theme.typography.body1,
-    }
-  };
-}
+const styles = (theme) => ({
+  root: {
+    ...theme.typography.body1,
+  }
+});
 
 class LogViewer extends Component {
-  constructor (props) {
+  constructor(props) {
     super(props);
 
     this.renderLogs = this.renderLogs.bind(this);
 
     this.eventView = React.createRef();
   }
-  componentDidMount () {
+
+  componentDidMount() {
     raf(this.renderLogs);
   }
-  renderLogs () {
+
+  renderLogs() {
     if (this.eventView.current && this.eventView.current.parentElement) {
-      let offset = TimelineWorker.currentOffset();
-      var lastEvent = TimelineWorker.lastEvents(50, offset);
+      const offset = TimelineWorker.currentOffset();
+      const lastEvent = TimelineWorker.lastEvents(50, offset);
       if (!lastEvent.length) {
         this.eventView.current.innerHTML = '';
       } else if (this.lastLastEvent !== lastEvent[0].LogMonoTime) {
@@ -50,12 +50,15 @@ class LogViewer extends Component {
       raf(this.renderLogs);
     }
   }
-  render () {
-    const propsToRender = {...this.props};
+
+  render() {
+    const propsToRender = { ...this.props };
     delete propsToRender.classes;
 
+    const { classes } = this.props;
+
     return (
-      <div className={ this.props.classes.root } style={{width: '100%', overflow: 'hidden'}} >
+      <div className={classes.root} style={{ width: '100%', overflow: 'hidden' }}>
         <pre ref={this.eventView} />
         <pre>{ JSON.stringify(propsToRender, null, 2) }</pre>
       </div>
@@ -63,8 +66,12 @@ class LogViewer extends Component {
   }
 }
 
-export default connect(mapStateToProps)(withStyles(styles)(LogViewer));
+LogViewer.propTypes = {
+  classes: PropTypes.object.isRequired,
+};
 
 function mapStateToProps(state) {
   return state.workerState;
 }
+
+export default connect(mapStateToProps)(withStyles(styles)(LogViewer));
