@@ -4,10 +4,11 @@ import ReactDOM from 'react-dom';
 import document from 'global/document';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { MuiThemeProvider } from '@material-ui/core/styles';
-import MuiPickersUtilsProvider from 'material-ui-pickers/utils/MuiPickersUtilsProvider';
+import MuiPickersUtilsProvider from 'material-ui-pickers/MuiPickersUtilsProvider';
 import DateFnsUtils from 'material-ui-pickers/utils/date-fns-utils';
 
 import './index.css';
+import 'react-virtualized/styles.css';
 import App from './App';
 import Theme from './theme';
 import { unregister } from './registerServiceWorker';
