@@ -1,4 +1,19 @@
 
+function getDefaultStartDate() {
+  const d = new Date();
+  d.setHours(d.getHours() + 1, 0, 0, 0);
+
+  return (new Date(d.getTime() - 1000 * 60 * 60 * 24 * 14)).getTime();
+}
+
+function getDefaultEndDate() {
+  const d = new Date();
+  d.setHours(d.getHours() + 1, 0, 0, 0);
+
+  // return (new Date(d.getTime() - 1000 * 60 * 60 * 36)).getTime();
+  return d.getTime();
+}
+
 module.exports = {
   start: getDefaultStartDate(),
   end: getDefaultEndDate(),
@@ -12,6 +27,10 @@ module.exports = {
   segment: 0,
   nextSegment: null,
   playSpeed: 0, // 0 = stopped, 1 = playing, 2 = 2x speed... multiplier on speed
+  desiredPlaySpeed: 0, // the speed the user has selected so that buffering can resume
+  shouldBuffer: true, // if we should bother waiting for buffered data or just roll on
+  bufferingVideo: false, // if we're currently buffering for more data
+  bufferingData: false, // if we're currently buffering for more data
   offset: 0, // in miliseconds from the start
   startTime: Date.now(), // millisecond timestamp in which play began
 
@@ -39,21 +58,7 @@ module.exports = {
     duration: null
   },
 
-  profile: null,
+  profile: {},
+  devices: [],
   isDemo: false,
 };
-
-function getDefaultStartDate () {
-  var d = new Date();
-  d.setHours(d.getHours() + 1, 0, 0, 0);
-
-  return (new Date(d.getTime() - 1000 * 60 * 60 * 24 * 14)).getTime();
-}
-
-function getDefaultEndDate () {
-  var d = new Date();
-  d.setHours(d.getHours() + 1, 0, 0, 0);
-
-  // return (new Date(d.getTime() - 1000 * 60 * 60 * 36)).getTime();
-  return d.getTime();
-}
