@@ -1,32 +1,36 @@
 const Event = require('geval/event');
-const LogReaderSharedWorker = require('./logReader.sharedworker');
 const LogReaderWorker = require('./logReader.worker');
+// const LogReaderSharedWorker = require('./logReader.sharedworker');
 
 const DataEvent = Event();
 
-module.exports = getWorker;
+let logReader = null;
 
-getWorker.onData = DataEvent.listen;
+function handleMessage(msg) {
+  DataEvent.broadcast(msg);
+}
 
-var logReader = null;
+function relayMessage(msg) {
+  const port = logReader.port || logReader;
+  port.postMessage(msg.data);
+}
 
-function getWorker () {
+function getWorker() {
   if (logReader) {
     return logReader;
   }
-  if (false && typeof LogReaderSharedWorker === 'function' && typeof SharedWorker === 'function') {
-    logReader = new LogReaderSharedWorker();
-  } else if (typeof LogReaderWorker === 'function') {
-    console.warn('Using web worker fallback');
+  // if (typeof LogReaderSharedWorker === 'function' && typeof SharedWorker === 'function') {
+  // logReader = new LogReaderSharedWorker();
+  if (typeof LogReaderWorker === 'function') {
     logReader = new LogReaderWorker();
   } else {
     throw new Error('Don\'t');
   }
-  let port = logReader.port || logReader;
+  const port = logReader.port || logReader;
 
-  let channel = new MessageChannel();
-  let logPort = channel.port1;
-  let apiPort = channel.port2;
+  const channel = new MessageChannel();
+  const logPort = channel.port1;
+  const apiPort = channel.port2;
 
   logPort.onmessage = relayMessage;
   port.onmessage = handleMessage;
@@ -34,11 +38,6 @@ function getWorker () {
   return apiPort;
 }
 
-function handleMessage (msg) {
-  DataEvent.broadcast(msg);
-}
+module.exports = getWorker;
 
-function relayMessage (msg) {
-  let port = logReader.port || logReader;
-  port.postMessage(msg.data);
-}
+getWorker.onData = DataEvent.listen;
