@@ -1,10 +1,10 @@
 import { createStore } from 'redux';
 import reduceReducers from 'reduce-reducers';
 
-import Playback from './playback';
+import { reducer as playbackReducer } from './playback';
 import Segments from './segments';
-import Reducers from './reducers'
+import Reducers from './reducers';
 
-const store = createStore(reduceReducers(Playback.reducer, Segments.reducer, Reducers.globalState));
+const store = createStore(reduceReducers(playbackReducer, Segments.reducer, Reducers.globalState));
 
 export default store;
